Add Camera * src/gfx/camera.hpp * src/gfx/camera.cpp * Panning and zooming also implemented

+++ src/ui.hpp
@@ -27,7 +27,7 @@ struct ContextState
 {
     tile::TileMap* tileMap = nullptr;
     tile::Tile currentTile = 0;
-    int tileSetIndex = 0;
+    uint16_t tileSetIndex = 0;
     bool hasMouseFocus = false;
     bool hasKeyboardFocus = false;
 }; // struct ContextState
